[STORE] Fixed, that `options` have not been passed to gateway in MyModel#update
In the `Rails` module, the `update` method has been redefined to allow
converting dates coming in from the Rails' forms, but the `options`
argument has not been passed down to the `update` method in the `Store` module.

This patch fixes that and adds unit and integration tests to cover the beahviour.

Closes #347